Refractor: implementations for `unpublish` were incorrect and were actually for `end experiment`. Renamed files and adjusted the button clicks to fit `End Experiment` option instead
